Fix compliance controls 500: replace Decimal score with float
Pydantic v2 serializes Decimal as string in JSON mode, causing
FastAPI response_model validation to fail with 500. Changed
score field to float in both schema and service.

+++ govguard/backend/modules/compliance_engine/schemas.py
@@ -1,6 +1,5 @@
 """GovGuard™ — Compliance Schemas"""
 from datetime import datetime
-from decimal import Decimal
 from typing import Optional
 from uuid import UUID
 from pydantic import BaseModel
@@ -22,7 +21,7 @@ class ControlResponse(BaseModel):
 
 class ControlListResponse(BaseModel):
     controls: list[ControlResponse]
-    score: Decimal
+    score: float
     total: int
     passing: int
     failing: int
